requirements.txt
discord.py~=2.3.2
python-dotenv~=0.21.0
python-aliexpress-api~=3.0.0
